+++ internal/bloblang/parser/dot_env_parser.go
@@ -1,9 +1,16 @@
 package parser
 
+import (
+	"fmt"
+	"os"
+)
+
 func dotEnvParser() Func {
 	assignmentParser := Sequence(
 		NotInSet('=', ' ', '\n', '#'),
+		Optional(SpacesAndTabs()),
 		Char('='),
+		Optional(SpacesAndTabs()),
 		Optional(OneOf(QuotedString(), NotInSet('#', ' ', '\n'))),
 		Optional(SpacesAndTabs()),
 	)
@@ -30,9 +37,9 @@ func dotEnvParser() Func {
 		vars := map[string]string{}
 		for _, line := range res.Payload.(DelimitedResult).Primary {
 			sequence, _ := line.([]interface{})
-			if len(sequence) == 4 {
+			if len(sequence) == 6 {
 				key := sequence[0].(string)
-				value, _ := sequence[2].(string)
+				value, _ := sequence[4].(string)
 				vars[key] = value
 			}
 		}
@@ -41,12 +48,18 @@ func dotEnvParser() Func {
 	}
 }
 
-// ParseDotEnv attempts to parse a .env file containing environment variable
+// ParseDotEnvFile attempts to parse a .env file containing environment variable
 // assignments, and returns either a map of key/value assignments or an error.
-func ParseDotEnv(input string) (map[string]string, error) {
+func ParseDotEnvFile(path string) (map[string]string, error) {
+	dotEnvBytes, err := os.ReadFile(path)
+	if err != nil {
+		return nil, err
+	}
+	input := string(dotEnvBytes)
 	res := dotEnvParser()([]rune(input))
 	if res.Err != nil {
-		return nil, res.Err
+		line, _ := LineAndColOf([]rune(input), res.Err.Input)
+		return nil, fmt.Errorf("%v: %v", line, res.Err.ErrorAtPositionStructured("", []rune(input)))
 	}
 	return res.Payload.(map[string]string), nil
 }